mdz/pytorch/CAM++/3_deploy/modelzoo/CAM++/requirements.txt
numpy>=1.19.2
scipy>=1.6.3
librosa>=0.9.1
soundfile>=0.12.1
soundcard>=0.4.2
resampy>=0.2.2
numba>=0.53.0
pydub~=0.25.1
matplotlib>=3.5.2
termcolor>=1.1.0
pillow>=8.3.2
tqdm>=4.64.1
visualdl>=2.2.3
pyyaml>=5.4.1
scikit-learn>=1.0.2
torchinfo>=1.7.2
av>=10.0.0
torchaudio >= 0.9.0
faiss-cpu >=1.7.0